Move benchmark components
